requirements.txt
pandas~=1.4.1
selenium~=4.1.0
webdriver-manager
mysql-connector
gcloud